add core migrations, refactor vue user component

+++ database/migrations/2019_12_09_082630_create_role_user_table.php
@@ -15,7 +15,7 @@ class CreateRoleUserTable extends Migration
     {
         Schema::create('role_user', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('role_id')->unsigned();
+            $table->bigInteger('role_id')->unsigned();
             $table->integer('user_id')->unsigned();
             $table->timestamps();
         });
